Move NUSF tests to test_sdoe.py
